requirements.txt
streamlit>=1.40.2
requests
firecrawl-py>=4.6.0
elevenlabs>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
langchain-google-genai
gtts